Pass entire arguments, instead of just bits.
This makes download_images more readable, but means that any
change in the arguments datastructure. Such as even a simple
renaming of --is_new will neccesitate a refactoring.

I've gone with this approach, because any future refactoring
caused will be less harmful than the increased readability
and simplicity gained from this change.